Mapping relationship change to dropdown menu on click
