+++ src/components/Projects/Tables/TableStyling.scss
@@ -98,7 +98,7 @@ input::file-selector-button:hover {
   border: 1px solid $greyColor;
   border-radius: 4px;
   &:hover {
-    background-color: rgb(239, 235, 235);
+    background-color: rgb(245, 245, 245);
     transition: all 0.3s linear;
   }
 }
