+++ src/components/Projects/Terminologies/EditMappingModal.jsx
@@ -19,6 +19,7 @@ import { SearchContext } from '../../../Contexts/SearchContext';
 import { OntologyFilterCodeSubmitTerm } from '../../Manager/MappingsFunctions/OntologyFilterCodeSubmitTerm';
 import { MappingRelationship } from '../../Manager/MappingsFunctions/MappingRelationship';
 import { MappingContext } from '../../../Contexts/MappingContext';
+import { EditMappingsLabel } from '../Tables/EditMappingsLabel';
 
 export const EditMappingsModal = ({
   editMappings,
@@ -32,7 +33,7 @@ export const EditMappingsModal = ({
   const [termMappings, setTermMappings] = useState([]);
   const [options, setOptions] = useState([]);
   const { vocabUrl, setSelectedKey, user } = useContext(myContext);
-  const { setShowOptions, showOptions, idsForSelect } =
+  const { setShowOptions, showOptions, relationshipOptions, idsForSelect } =
     useContext(MappingContext);
   const {
     apiPreferencesCode,
@@ -55,6 +56,7 @@ export const EditMappingsModal = ({
     setOptions([]);
     setSelectedKey(null);
     setApiPreferencesCode(undefined);
+    setShowOptions(false);
   };
 
   const fetchMappings = () => {
@@ -104,7 +106,10 @@ export const EditMappingsModal = ({
             mappings.push(val); // For each mapping in the mappings array, push the stringified object above to the mappings array.
             // For each mapping in the mappings array, push the stringified object above to the options array
             // as the value for the value field for the ant.design checkbox. The label for the checkbox is returned in edditMappingsLabel function.
-            options.push({ value: val, label: editMappingsLabel(m, index) });
+            options.push({
+              value: val,
+              label: <EditMappingsLabel item={m} index={index} />,
+            });
           });
           // termMappings are set to the mappings array. Options are set to the options array.
           setTermMappings(mappings);
@@ -122,54 +127,6 @@ export const EditMappingsModal = ({
         .finally(() => setLoading(false));
     }
   };
-
-  // The label for the checkbox for each mapping. Displays JSX to show the display and code.
-  // The description field is commented out to be integrated when there is API support.
-  const editMappingsLabel = (item, index) => {
-    return (
-      <>
-        <div key={index} className="modal_search_result">
-          <div>
-            <div className="modal_term_ontology">
-              <div>
-                <b>{item?.display}</b>
-              </div>
-              <div>{item?.code}</div>
-              <div
-              // className="mapping_relationship"
-              // onClick={setShowOptions(item.mapping_relationship)}
-              >
-                <MappingRelationship mapping={item} />
-
-                {/* {item?.mapping_relationship && !showOptions
-                  ? displayRelationship(item)
-                  : (!item.mapping_relationship || !!showOptions) && (
-                      <MappingRelationship mapping={item} />
-                    )} */}
-              </div>
-            </div>
-            <div>
-              {item?.description?.length > 100 ? (
-                <Tooltip
-                  title={item?.description}
-                  placement="topRight"
-                  mouseEnterDelay={0.5}
-                >
-                  {ellipsisString(item?.description, '100')}
-                </Tooltip>
-              ) : (
-                ellipsisString(item?.description, '100')
-              )}
-            </div>
-          </div>
-        </div>
-      </>
-    );
-  };
-
-  const searchProp = editMappings?.display
-    ? editMappings.display
-    : editMappings?.code;
 
   // Function to send a PUT call to update the mappings.
   // Each mapping in the mappings array being edited is JSON.parsed and mappings are turned into objects in the mappings array.
